requirements.txt
beautifulsoup4>=4.13.4
fastapi>=0.115.13
uvicorn>=0.34.3
pydantic>=2.0.0
httpx>=0.27.0
reportlab>=4.0.0
playwright>=1.52.0
psycopg2-binary>=2.9.10
pypdf2>=3.0.1
python-docx>=1.2.0
python-multipart>=0.0.20
requests>=2.32.4
selenium>=4.33.0
jinja2>=3.1.6
email-validator>=2.2.0